De-linting for gcc. Makefile modifications.

+++ g_utils.c
@@ -373,7 +373,7 @@ char *G_CopyString (char *in)
 {
 	char	*out;
 	
-	out = gi.TagMalloc (strlen(in)+1, TAG_LEVEL);
+	out = gi.TagMalloc ((int)strlen(in)+1, TAG_LEVEL);
 	strlcpy (out, in, strlen(in)+1);
 	return out;
 }
